% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-NullObject.R
\name{NullObject}
\alias{NullObject}
\title{Null Object Pattern}
\usage{
NullObject()
}
\description{
Model a domain concept using natural lingo of the domain
experts, such as “Passenger”, “Address”, and “Money”.
}
\examples{
# See more examples at <https://tidylab.github.io/R6P/articles>

colnames(NullObject())
nrow(NullObject())
}
\seealso{
Other base design patterns: 
\code{\link{Singleton}},
\code{\link{ValueObject}()}
}
\concept{base design patterns}
