% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6P-package.R
\docType{package}
\name{R6P-package}
\alias{R6P}
\alias{R6P-package}
\title{R6P: Design Patterns in R}
\description{
Build robust and maintainable software with object-oriented design patterns in R. Design patterns abstract and present in neat, well-defined components and interfaces the experience of many software designers and architects over many years of solving similar problems. These are solutions that have withstood the test of time with respect to re-usability, flexibility, and maintainability. 'R6P' provides abstract base classes with examples for a few known design patterns. The patterns were selected by their applicability to analytic projects in R. Using these patterns in R projects have proven effective in dealing with the complexity that data-driven applications possess.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tidylab.github.io/R6P/}
  \item \url{https://github.com/tidylab/R6P}
  \item Report bugs at \url{https://github.com/tidylab/R6P/issues}
}

}
\author{
\strong{Maintainer}: Harel Lustiger \email{tidylab@gmail.com} (\href{https://orcid.org/0000-0003-2953-9598}{ORCID})

Other contributors:
\itemize{
  \item Tidylab [copyright holder, funder]
}

}
\keyword{internal}
